refactor(shaper): keep verbosity steering, drop every other output lever
Measured on claude-sonnet-5 against the live API rather than reasoned about,
which is what changed the design.

WHAT THE MEASUREMENTS SAID

Effort does move output, hard, on the turns where thinking is the work:

    effort=low    median  41 tokens   [41, 50, 41]
    effort=high   median 668 tokens   [739, 657, 668]

Both answers were correct (216 miles); the extra 627 tokens were headers,
a restated setup and an explanation of the trick. That waste is real and it
is what steering targets.

But the router lowered effort on MECHANICAL tool-result continuations, and
those turns barely emit anything to save -- 58 tokens at high, 9 at low, so
~$0.0007. A growing-conversation probe put the cost of one effort switch at
one extra turn-delta of cache re-write, ~$0.011 at a 3k delta. Roughly 15x
underwater, and structurally so: the lever fired only deep in a session,
where the delta is largest. The savings that do exist live on reasoning
turns, which are exactly the turns not to cheapen.

Two further findings retired the rest:

  * `thinking.budget_tokens` is rejected outright now --
    "thinking.type.enabled is not supported for this model. Use
    thinking.type.adaptive and output_config.effort" -- so that lever was
    dead code on every current model.
  * `max_tokens` provably does NOT invalidate the cache (512->2048 read
    clean), but a ceiling only saves by truncating, which on an agent turn
    can sever a tool call mid-JSON. `text.verbosity` was never measured.

Also worth recording, since it corrects an assumption in the removed code:
each effort value keeps its OWN cache entry. Alternating between two warm
values reads at full and writes nothing, so switching is not the catastrophe
the earlier design took it for. It simply does not pay for what it buys.

WHAT IS LEFT

Verbosity steering, and nothing else. ~60 cached tokens, set once per
conversation so the prefix is never disturbed, cannot truncate, works on
every model and every wire format.

REMOVED

  route_effort / route_openai_reasoning_effort / route_openai_text_verbosity
  / route_responses_effort, output_effort_policy.py, the effort_router_enabled
  and mechanical_effort settings, HEADROOM_EFFORT_ROUTER and
  HEADROOM_MECHANICAL_EFFORT, and the agent_savings profile flag.

  plugins/headroom-output entirely. It was built to host a rich multi-lever
  system; with only steering surviving, ~2,500 of its 3,539 lines were
  machinery for levers that no longer exist, and the core already applies
  steering on all four wire formats and already carries the whole
  measurement path (baseline, holdout, stratification, CIs) in
  output_savings.py.

  PipelineStage.PRE_SEND_PARAMS and PipelineEvent.body, the seam that let an
  extension rewrite outbound parameters. Nothing writes to it now, and an
  unused writable-body hook on live traffic is a liability, not an extension
  point. OUTCOME_OBSERVED stays: it is how an extension measures its effect.

Kept deliberately: the thinking/visible output-token split, which is how the
one surviving lever gets measured at all.

Full suite: 23 failures, identical to the pre-change baseline. Net -4,392 lines.
