--- plugins/headroom-output/.gitignore
@@ -1,6 +1,0 @@
-.venv/
-__pycache__/
-*.pyc
-.ruff_cache/
-.pytest_cache/
-*.egg-info/
