--- plugins/headroom-output/pyproject.toml
@@ -1,69 +1,0 @@
-[build-system]
-requires = ["setuptools>=61"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "headroom-output"
-version = "0.1.0"
-description = "Output-token shaping levers for Headroom: effort routing, adaptive ceilings, verbosity steering"
-readme = "README.md"
-requires-python = ">=3.10"
-license = { text = "Apache-2.0" }
-authors = [{ name = "Tejas Chopra" }]
-keywords = ["headroom", "llm", "output-tokens", "cost", "proxy", "shaping"]
-# Standard library only. The levers are pure functions over plain data and the
-# ledgers are JSON — a plugin that sits in the proxy request path should not be
-# dragging a dependency tree into it.
-dependencies = []
-classifiers = [
-    "Development Status :: 3 - Alpha",
-    "Intended Audience :: Developers",
-    "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Topic :: Internet :: Proxy Servers",
-]
-
-[project.optional-dependencies]
-dev = ["pytest>=7", "ruff"]
-test = ["pytest>=7"]
-
-[project.urls]
-Homepage = "https://github.com/chopratejas/headroom"
-Repository = "https://github.com/chopratejas/headroom"
-Issues = "https://github.com/chopratejas/headroom/issues"
-
-# Adds the `headroom output` command group. This seam exists today and needs no
-# core change, so the plugin is inspectable from the CLI the moment it is
-# installed — before the request-path seam it ultimately wants is built.
-[project.entry-points."headroom.cli_extension"]
-output = "headroom_output.cli:register"
-
-# The plugin defines its own seam. A third party (or an enterprise with a house
-# answer format) adds a lever by declaring one of these, without forking the
-# plugin — the same discovery-then-explicit-selection model the core uses for
-# compressors and proxy extensions.
-[project.entry-points."headroom.output_lever"]
-verbosity = "headroom_output.levers.steering:VerbosityLever"
-effort = "headroom_output.levers.effort:EffortLever"
-thinking_budget = "headroom_output.levers.thinking:ThinkingBudgetLever"
-text_verbosity = "headroom_output.levers.text_verbosity:TextVerbosityLever"
-max_tokens = "headroom_output.levers.max_tokens:AdaptiveMaxTokensLever"
-
-[tool.setuptools.packages.find]
-where = ["src"]
-
-[tool.ruff]
-line-length = 100
-target-version = "py310"
-
-# Ruff does not merge a parent config, and an empty [tool.ruff] here would
-# shadow the repo root's rule set entirely — linting this package against a
-# different set of rules than the rest of the tree. Mirror the root's select.
-[tool.ruff.lint]
-select = ["E", "W", "F", "I", "B", "C4", "UP"]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
