+++ tests/test_openai_codex_ws_lifecycle.py
@@ -412,7 +412,8 @@ async def test_ws_first_frame_output_shaper_rewrites_without_compression(monkeyp
     sent = json.loads(upstream.sent[0])
     payload = sent["response"]
     assert "<headroom_output_shaping>" in payload["instructions"]
-    assert payload["text"]["verbosity"] == "low"
+    # text.verbosity is no longer injected: steering is the only lever.
+    assert "text" not in payload
     assert any(t == "output_shaper:verbosity:L2" for t in outcomes[-1].transforms_applied)
